improvement(delete-service): Safely delete NewsService. Rename NewsRepository to NewsRepositoryImpl. Create NewsRepository interface so it can be implemented within the repository.
