+++ app/src/main/java/com/enriqueajin/newsapp/data/NewsRepositoryImpl.kt
@@ -3,16 +3,17 @@ package com.enriqueajin.newsapp.data
 import androidx.paging.Pager
 import androidx.paging.PagingConfig
 import androidx.paging.PagingData
-import com.enriqueajin.newsapp.data.network.NewsService
+import com.enriqueajin.newsapp.data.network.NewsApiClient
 import com.enriqueajin.newsapp.data.network.model.NewsItem
+import com.enriqueajin.newsapp.domain.repository.NewsRepository
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
 import javax.inject.Singleton
 
 @Singleton
-class NewsRepository @Inject constructor(private val newsService: NewsService) {
+class NewsRepositoryImpl @Inject constructor(private val api: NewsApiClient): NewsRepository {
 
-    fun getNews(
+    override fun getNews(
         category: String,
         pageSize: Int,
         needsPagination: Boolean
@@ -20,14 +21,14 @@ class NewsRepository @Inject constructor(private val newsService: NewsService) {
         return Pager(config = PagingConfig(pageSize = pageSize, prefetchDistance = PREFETCH_ITEMS),
             pagingSourceFactory = {
                 NewsPagingSource(
-                    service = newsService,
+                    api = api,
                     category = category,
                     needsPagination = needsPagination
                 )
             }).flow
     }
 
-    fun getNewsByKeyword(
+    override fun getNewsByKeyword(
         keyword: String,
         pageSize: Int,
         needsPagination: Boolean
@@ -35,7 +36,7 @@ class NewsRepository @Inject constructor(private val newsService: NewsService) {
         return Pager(config = PagingConfig(pageSize = pageSize, prefetchDistance = PREFETCH_ITEMS),
             pagingSourceFactory = {
                 NewsPagingSource(
-                    service = newsService,
+                    api = api,
                     keyword = keyword,
                     needsPagination = needsPagination
                 )
