+++ app/src/main/java/com/enriqueajin/newsapp/domain/use_case/GetNewsByCategoryUseCase.kt
@@ -1,15 +1,15 @@
-package com.enriqueajin.newsapp.domain
+package com.enriqueajin.newsapp.domain.use_case
 
 import androidx.paging.PagingData
-import com.enriqueajin.newsapp.data.NewsRepository
+import com.enriqueajin.newsapp.data.NewsRepositoryImpl
 import com.enriqueajin.newsapp.data.network.model.NewsItem
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
 
 class GetNewsByCategoryUseCase @Inject constructor(
-    private val newsRepository: NewsRepository
+    private val newsRepositoryImpl: NewsRepositoryImpl
 ) {
     operator fun invoke(category: String = "general", pageSize: Int, needsPagination: Boolean = false):Flow<PagingData<NewsItem>> {
-        return newsRepository.getNews(category, pageSize, needsPagination)
+        return newsRepositoryImpl.getNews(category, pageSize, needsPagination)
     }
 }
