+++ app/src/main/java/com/enriqueajin/newsapp/ui/keyword_news/KeywordNewsViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.viewModelScope
 import androidx.paging.PagingData
 import androidx.paging.cachedIn
 import com.enriqueajin.newsapp.data.network.model.NewsItem
-import com.enriqueajin.newsapp.domain.GetNewsByKeywordUseCase
+import com.enriqueajin.newsapp.domain.use_case.GetNewsByKeywordUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
